requirements.txt
# Core LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchainhub>=0.1.0
langgraph>=0.2.0
langserve>=0.3.0

# OpenAI & tokenization
openai>=1.0.0
tiktoken>=0.7.0

# Vector store / data handling
chromadb>=0.5.0
pypdf>=4.2.0
unstructured>=0.14.0

# Web / API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
sse-starlette>=3.0.0

# Env & utils
python-dotenv>=1.0.1
